requirements.txt
aws-cdk-lib
aws-cdk.lambda-layer-kubectl-v29
boto3
checkov
constructs
flake8
gitpython
pylint
pytest
pyyaml
requests
semantic-version
